+++ CHANGELOG.md
@@ -19,7 +19,7 @@ So we can convert the units correctly
 * WireMixin: abscissa (add tolerance as parameter)
 * OpenRoundedLineSegment2D: deleted discretization_points() so it uses the one from WireMixin.
 * Contour2D: moved bounding_rectangle and get_bounding_rectangle to Wire2D. 
-
+* BSplineCurve: from_points_interpolation, uses centripedal method for better fitting.
 ### Removed